move proof-of-work logic in its own module
