+++ apps/blockchain/lib/blockchain/chain.ex
@@ -6,7 +6,7 @@ defmodule Blockchain.Chain do
 
   use GenServer
 
-  alias Blockchain.{Block, Data}
+  alias Blockchain.{Block, Data, ProofOfWork}
 
   def start_link do
     GenServer.start_link(__MODULE__, nil, name: __MODULE__)
@@ -68,7 +68,7 @@ defmodule Blockchain.Chain do
         {:error, "invalid index"}
       previous_block.hash != block.previous_hash ->
         {:error, "invalid previous hash"}
-      Block.verify_proof_of_work(block.hash) == false ->
+      ProofOfWork.verify(block.hash) == false ->
         {:error, "no proof of work"}
       block.hash != Block.compute_hash(block) ->
         {:error, "invalid block hash"}
